consolidate QueueManager.jsx and queue/index.jsx; extract QueueManager from BaseContainer
