+++ app/views/queue/index.html.erb
@@ -1,8 +1,7 @@
 <% content_for :full_page_content do %>
-  <%= react_component("BaseContainer", props: {
+  <%= react_component("QueueManager", props: {
     userDisplayName: current_user.display_name,
     dropdownUrls: dropdown_urls,
-    page: "WorkQueue",
     feedbackUrl: feedback_url,
     buildDate: build_date
   }) %>
